+++ packages/swr-login/CHANGELOG.md
@@ -1,5 +1,13 @@
 # swr-login
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @swr-login/core@0.6.0
+  - @swr-login/react@0.6.0
+
 ## 0.2.4
 
 ### Patch Changes